feat(logger): allow controlling level with env var (#452)
