+++ garden-service/test/logger/log-node.ts
@@ -1,0 +1,31 @@
+import { expect } from "chai"
+import { getLogger } from "../../src/logger/logger"
+
+const logger = getLogger()
+
+beforeEach(() => {
+  (<any>logger).children = []
+})
+
+describe("LogNode", () => {
+  describe("appendNode", () => {
+    it("should add new child entries to the respective node", () => {
+      logger.error("error")
+      logger.warn("warn")
+      logger.info("info")
+      logger.verbose("verbose")
+      logger.debug("debug")
+      logger.silly("silly")
+
+      const prevLength = logger.children.length
+      const entry = logger.children[0]
+      const nested = entry.info("nested")
+      const deepNested = nested.info("deep")
+
+      expect(logger.children[0].children).to.have.lengthOf(1)
+      expect(logger.children[0].children[0]).to.eql(nested)
+      expect(logger.children[0].children[0].children[0]).to.eql(deepNested)
+      expect(logger.children).to.have.lengthOf(prevLength)
+    })
+  })
+})
