+++ garden-service/test/logger/logger.ts
@@ -1,0 +1,60 @@
+import { expect } from "chai"
+
+import { LogLevel } from "../../src/logger/log-node"
+import { getLogger } from "../../src/logger/logger"
+
+const logger = getLogger()
+
+beforeEach(() => {
+  (<any>logger).children = []
+})
+
+describe("Logger", () => {
+  describe("findById", () => {
+    it("should return the first log entry with a matching id and undefined otherwise", () => {
+      logger.info({ msg: "0" })
+      logger.info({ msg: "a1", id: "a" })
+      logger.info({ msg: "a2", id: "a" })
+      expect(logger.findById("a")["opts"]["msg"]).to.eql("a1")
+      expect(logger.findById("z")).to.be.undefined
+    })
+  })
+
+  describe("filterBySection", () => {
+    it("should return an array of all entries with the matching section name", () => {
+      logger.info({ section: "s0" })
+      logger.info({ section: "s1", id: "a" })
+      logger.info({ section: "s2" })
+      logger.info({ section: "s1", id: "b" })
+      const s1 = logger.filterBySection("s1")
+      const sEmpty = logger.filterBySection("s99")
+      expect(s1.map(entry => entry.id)).to.eql(["a", "b"])
+      expect(sEmpty).to.eql([])
+    })
+  })
+
+  describe("getLogEntries", () => {
+    it("should return an ordered list of log entries", () => {
+      logger.error("error")
+      logger.warn("warn")
+      logger.info("info")
+      logger.verbose("verbose")
+      logger.debug("debug")
+      logger.silly("silly")
+
+      const entries = logger.getLogEntries()
+      const levels = entries.map(e => e.level)
+
+      expect(entries).to.have.lengthOf(6)
+      expect(levels).to.eql([
+        LogLevel.error,
+        LogLevel.warn,
+        LogLevel.info,
+        LogLevel.verbose,
+        LogLevel.debug,
+        LogLevel.silly,
+      ])
+    })
+  })
+
+})
